+++ src/path.rs
@@ -29,20 +29,7 @@ pub enum PathOp {
     PenWidth(f32),
 }
 
-/// Path2D is a container for arbitrary path operations.
-///
-/// Use [PathBuilder](struct.PathBuilder.html) to construct paths.
-pub struct Path2D {
-    ops: Vec<PathOp>,
-}
-
-/// IterPath2D is an iterator for Path2D structs.
-pub struct IterPath2D<'a> {
-    path: &'a Path2D,
-    pos: usize,
-}
-
-/// Builder for [Path2D](struct.Path2D.html).
+/// Path Builder.
 ///
 /// # Example
 /// ```
@@ -60,32 +47,6 @@ pub struct PathBuilder {
     absolute: bool,
     /// Current pen position
     pen: Pt,
-}
-
-impl Path2D {
-    /// Get an iterator of path operations.
-    pub fn iter(&self) -> IterPath2D {
-        IterPath2D { path: self, pos: 0 }
-    }
-}
-
-impl<'a> IntoIterator for &'a Path2D {
-    type Item = &'a PathOp;
-    type IntoIter = IterPath2D<'a>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.iter()
-    }
-}
-
-impl<'a> Iterator for IterPath2D<'a> {
-    type Item = &'a PathOp;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let p = self.pos;
-        self.pos += 1;
-        self.path.ops.get(p)
-    }
 }
 
 impl Default for PathBuilder {
@@ -205,7 +166,7 @@ impl PathBuilder {
     }
 
     /// Build path from specified operations.
-    pub fn build(self) -> Path2D {
-        Path2D { ops: self.ops }
+    pub fn build(self) -> Vec<PathOp> {
+        self.ops
     }
 }